+++ SlyceSDK/SlyceSDK.framework/Versions/A/Headers/SFSlyce.h
@@ -34,12 +34,36 @@
 /*!
  *  @method
  *
- *  @brief Initializes a new SFSlyce object.
+ *  @brief Initializes a new SFSlyce object using valid App ID and App Key credentials.
  *
- *  @discussion Use this method to open the SDK for 2D/3D recognition functionality.
- *  Depending on your client ID, the SDK will be able to perform standard/premium tasks.
+ *  @discussion Use this method to open the SDK for 'Public' recognition functionality.
  *
- *  @param clientID the client ID.
+ *  @param appId the App ID.
+ *  @param appKey the App Key.
+ *  @param error the pointer to the error object, if any.
+ *
+ *  @return `YES` if it succeeded, `NO` otherwise.
+ *
+ *  @warning As this method performs validation checks of your client ID on our
+ *  servers, call this method off the main (UI) thread!
+ *
+ *  **Important:** Successfully opening the SFSlyce object is mandatory before you start
+ *  performing any tasks in the SDK!
+ *
+ */
+
+- (BOOL)openWithAppId:(NSString *)appId appKey:(NSString *)appKey error:(NSError **)error;
+
+/*!
+ *  @method
+ *
+ *  @brief Initializes a new SFSlyce object using a customer ID.
+ *
+ *  @discussion Use this method to open the SDK for 2D/3D 'Premium' recognition functionality.
+ *  Depending on your client ID, the SDK will be able to perform Standard and/or Premium  
+ *  tasks.
+ *
+ *  @param clientID the customer ID.
  *  @param error the pointer to the error object, if any.
  *
  *  @return `YES` if it succeeded, `NO` otherwise.
@@ -62,7 +86,27 @@
  
  @property
  
- @abstract The client ID used to initialize the SFSlyce object.
+ @abstract The app ID used to initialize the SFSlyce object for 'Public' mode.
+ 
+ */
+
+@property (nonatomic, readonly, copy) NSString *appID;
+
+/*!
+ 
+ @property
+ 
+ @abstract The app key used to initialize the SFSlyce object for 'Public' mode.
+ 
+ */
+
+@property (nonatomic, readonly, copy) NSString *appKey;
+
+/*!
+ 
+ @property
+ 
+ @abstract The client ID used to initialize the SFSlyce object for 'Premium' mode.
  
  */
 
@@ -93,6 +137,19 @@
  */
 
 @property (nonatomic, readonly) BOOL is2DSearchEnabled;
+
+/*!
+ 
+ @property
+ 
+ @abstract Indicates whether the user can crop the image after snapping.
+ @discussion default is `NO`.
+ 
+ @note The value of this property depends on your client ID.
+ 
+ */
+
+@property (nonatomic, readonly) BOOL isCroppingEnabled;
 
 /*!