+++ SlyceSDKDemo/SlyceSDKDemo/MyViewController.m
@@ -12,34 +12,95 @@
 #import "MBProgressHUD.h"
 #import <SlyceSDK/SlyceSDK.h>
 
+
 #define kClientID @"ClientID"
+#define kAppID @"AppID"
+#define kAppKey @"AppKey"
+#define kClientMode @"ClientMode"
+
+typedef NS_ENUM(NSUInteger, ClientMode)
+{
+    ClientModePremium,
+    ClientModePublic
+};
 
 @interface MyViewController () <SFCameraViewControllerDelegate, SFRequestDelegate, UITextFieldDelegate, UIAlertViewDelegate, NSStreamDelegate, NSURLConnectionDataDelegate>
 
 @property (nonatomic, copy) NSArray *products;
 @property (nonatomic, strong) MBProgressHUD *hud;
 @property (weak, nonatomic) IBOutlet UITextField *clientIDTextField;
+@property (weak, nonatomic) IBOutlet UITextField *appIDTextField;
+@property (weak, nonatomic) IBOutlet UITextField *appKeyTextField;
+
 @property (copy, nonatomic) NSString *clientID;
+@property (copy, nonatomic) NSString *appID;
+@property (copy, nonatomic) NSString *appKey;
 @property (weak, nonatomic) IBOutlet UILabel *premiumLabel;
+@property (weak, nonatomic) IBOutlet UILabel *publicLabel;
 @property (weak, nonatomic) IBOutlet UILabel *versionLabel;
 @property (weak, nonatomic) IBOutlet UILabel *results2DLabel;
 @property (weak, nonatomic) IBOutlet UILabel *results2DAdditionalLabel;
+@property (weak, nonatomic) IBOutlet UIView *premiumView;
+@property (weak, nonatomic) IBOutlet UIView *publicView;
+@property (weak, nonatomic) IBOutlet UIButton *premiumPublicButton;
 
 @property (nonatomic, strong) SFSlyce *slyce;
 @property (nonatomic, strong) SFCameraViewController *cameraVC;
+
+@property (nonatomic) ClientMode clientMode;
 
 @end
 
 @implementation MyViewController
 
-@synthesize clientID = _clientID;
-
 #pragma mark -
 #pragma mark Accessors
 
+- (void)setClientMode:(ClientMode)clientMode
+{
+    _clientMode = clientMode;
+    
+    if (clientMode == ClientModePremium)
+    {
+        [_premiumPublicButton setTitle:@"PUBLIC" forState:UIControlStateNormal];
+        
+        [UIView animateWithDuration:0.2 animations:^{
+           
+            _premiumView.alpha = 1;
+            _publicView.alpha = 0;
+        }];
+        
+        self.clientID = [[NSUserDefaults standardUserDefaults] objectForKey:kClientID];
+        
+        if (self.clientID && ! [self.clientID isEqualToString:@""])
+            [self acceptButtonClicked:nil];
+        
+        [[NSUserDefaults standardUserDefaults] setObject:@"premium" forKey:kClientMode];
+        [[NSUserDefaults standardUserDefaults] synchronize];
+    }
+    else if (clientMode == ClientModePublic)
+    {
+        [_premiumPublicButton setTitle:@"PREMIUM" forState:UIControlStateNormal];
+        
+        [UIView animateWithDuration:0.2 animations:^{
+            
+            _premiumView.alpha = 0;
+            _publicView.alpha = 1;
+        }];
+        
+        self.appID = [[NSUserDefaults standardUserDefaults] objectForKey:kAppID];
+        self.appKey = [[NSUserDefaults standardUserDefaults] objectForKey:kAppKey];
+        
+        if (self.appID && ! [self.appID isEqualToString:@""] && self.appKey && ! [self.appKey isEqualToString:@""])
+            [self acceptButtonClicked:nil];
+        
+        [[NSUserDefaults standardUserDefaults] setObject:@"public" forKey:kClientMode];
+        [[NSUserDefaults standardUserDefaults] synchronize];
+    }
+}
+
 - (void)setClientID:(NSString *)clientID
 {
-    _clientID = clientID;
     _clientIDTextField.text = clientID;
     [[NSUserDefaults standardUserDefaults] setObject:clientID forKey:kClientID];
     [[NSUserDefaults standardUserDefaults] synchronize];
@@ -47,7 +108,31 @@
 
 - (NSString *)clientID
 {
-    return self.clientIDTextField.text;
+    return _clientIDTextField.text;
+}
+
+- (void)setAppID:(NSString *)appID
+{
+    _appIDTextField.text = appID;
+    [[NSUserDefaults standardUserDefaults] setObject:appID forKey:kAppID];
+    [[NSUserDefaults standardUserDefaults] synchronize];
+}
+
+- (NSString *)appID
+{
+    return _appIDTextField.text;
+}
+
+- (void)setAppKey:(NSString *)appKey
+{
+    _appKeyTextField.text = appKey;
+    [[NSUserDefaults standardUserDefaults] setObject:appKey forKey:kAppKey];
+    [[NSUserDefaults standardUserDefaults] synchronize];
+}
+
+- (NSString *)appKey
+{
+    return _appKeyTextField.text;
 }
 
 #pragma mark -
@@ -80,9 +165,24 @@
 #pragma mark -
 #pragma mark SFRequestDelegate
 
+- (void)sfRequest:(SFRequest *)sfRequest didStartForImage:(UIImage *)image
+{
+    NSLog(@"sfRequest:didStartForImage:");
+}
+
+- (void)sfRequest:(SFRequest *)sfRequest didStartForImageURL:(NSURL *)imageURL
+{
+    NSLog(@"sfRequest:didStartForImageURL:%@", imageURL.absoluteString);
+}
+
+- (void)sfRequest:(SFRequest *)sfRequest didFinishWithStatus:(SFStatusType)statusType
+{
+    NSLog(@"sfRequest:didFinishWithStatus:%lu", (unsigned long)statusType);
+}
+
 - (void)sfRequest:(SFRequest *)sfRequest didFailWithError:(NSError *)error
 {
-    NSLog(@"Error = %@", [error localizedDescription]);
+    NSLog(@"sfRequest:didFailWithError:%@", [error localizedDescription]);
     if (error.code != SFErrorCredMismatch)
         [self.hud hide:YES];
     
@@ -100,13 +200,13 @@
 //Premium
 - (void)sfRequest:(SFRequest *)sfRequest didDetectImage:(NSDictionary *)imageInfo
 {
-    NSLog(@"Recognized 2D Products: data = %@", imageInfo);
+    NSLog(@"sfRequest:didDetectImage:%@", imageInfo);
     self.results2DLabel.text = [imageInfo description];
 }
 
 - (void)sfRequest:(SFRequest *)sfRequest didReceiveImageInfo:(NSArray *)products
 {
-    NSLog(@"2D additional info = %@", products);
+    NSLog(@"sfRequest:didReceiveImageInfo:%@", products);
     self.products = [products copy];
     [self.hud hide:YES];
     [self performSegueWithIdentifier:@"ProductsSegue" sender:nil];
@@ -115,7 +215,7 @@
 - (void)sfRequest:(SFRequest *)sfRequest didReceiveResults:(NSDictionary *)results
 {
     NSArray *products = [results objectForKey:@"products"];
-    NSLog(@"Recognized 3D Products = %@", products);
+    NSLog(@"sfRequest:didReceiveResults:%@", products);
     [self.hud hide:YES];
     
     if (products && products.count > 0)
@@ -132,13 +232,13 @@
 
 - (void)sfRequest:(SFRequest *)sfRequest didFinishWithItemDescription:(NSDictionary *)itemDescription
 {
-    NSLog(@"Recognized 3D Item Description = %@", itemDescription);
+    NSLog(@"sfRequest:didFinishWithItemDescription:%@", itemDescription);
     [self.hud hide:YES];
 }
 
 - (void)sfRequest:(SFRequest *)sfRequest didFinishWithMerchantIDs:(NSArray *)merchantIDs
 {
-    NSLog(@"Supported Merchant IDs = %@", merchantIDs);
+    NSLog(@"sfRequest:didFinishWithMerchantIDs:%@", merchantIDs);
     [self.hud hide:YES];
 }
 
@@ -147,10 +247,10 @@
     switch (stage)
     {
         case SFRequestStageSendingImage:
-            NSLog(@"Current search stage = 'Sending Image'");
+            NSLog(@"sfRequest:didProgressToStage:'Sending Image'");
             break;
         case SFRequestStageAnalyzingImage:
-            NSLog(@"Current search stage = 'Analyzing Image'");
+            NSLog(@"sfRequest:didProgressToStage:'Analyzing Image'");
             break;
         default:
             break;
@@ -159,37 +259,52 @@
 
 - (void)sfRequest:(SFRequest *)sfRequest didProgressToValue:(CGFloat)value withMessage:(NSString *)message
 {
-    NSLog(@"Finished %.2f percents, current step = %@", value, message);
+    NSLog(@"sfRequest:didProgressToValue:%.2f withMessage:%@", value, message);
     self.hud.labelText = message;
 }
 
 #pragma mark -
 #pragma mark SFCameraViewControllerDelegate
 
+- (void)sfCameraViewController:(SFCameraViewController *)cameraViewController didFinishWithStatus:(SFStatusType)statusType
+{
+    NSLog(@"sfCameraViewController:didFinishWithStatus:%lu", (unsigned long)statusType);
+}
+
+- (void)sfCameraViewController:(SFCameraViewController *)cameraViewController didStartRequest:(SFRequest *)request forImage:(UIImage *)image
+{
+    NSLog(@"sfCameraViewController:didStartRequest:forImage:");
+}
+
+- (void)sfCameraViewController:(SFCameraViewController *)cameraViewController didSnapImage:(UIImage *)image
+{
+    NSLog(@"sfCameraViewController:didSnapImage:");
+}
+
 - (void)sfCameraViewControllerWasDismissed:(SFCameraViewController *)cameraViewController
 {
-    NSLog(@"SFCameraViewController was dismissed");
+    NSLog(@"sfCameraViewControllerWasDismissed:");
 }
 
 - (void)sfCameraViewController:(SFCameraViewController *)cameraViewController didProgressToValue:(CGFloat)value withMessage:(NSString *)message
 {
-    NSLog(@"Finished %.2f percents, current step = %@", value, message);
+    NSLog(@"sfCameraViewController:didProgressToValue:%.2f withMessage:%@", value, message);
 }
 
 - (void)sfCameraViewController:(SFCameraViewController *)cameraViewController didFailWithError:(NSError *)error
 {
     NSString *message = error.domain == SlyceErrorDomain ? [error sf_message] : [error localizedDescription];
-    NSLog(@"Error = %@", message);
+    NSLog(@"sfCameraViewController:didFailWithError:%@", message);
 }
 
 - (void)sfCameraViewController:(SFCameraViewController *)cameraViewController didDetectImage:(NSDictionary *)imageInfo
 {
-    NSLog(@"Recognized 2D Products: data = %@", imageInfo);
+    NSLog(@"sfCameraViewController:didDetectImage:%@", imageInfo);
 }
 
 - (void)sfCameraViewController:(SFCameraViewController *)cameraViewController didReceiveImageInfo:(NSArray *)products
 {
-    NSLog(@"Recognized 2D Products = %@", products);
+    NSLog(@"sfCameraViewController:didReceiveImageInfo:%@", products);
     self.products = products;
     
     if (products && products.count > 0)
@@ -199,7 +314,7 @@
 - (void)sfCameraViewController:(SFCameraViewController *)cameraViewController didReceiveResults:(NSDictionary *)results
 {
     NSArray *products = [results objectForKey:@"products"];
-    NSLog(@"Recognized 3D Products = %@", products);
+    NSLog(@"sfCameraViewController:didReceiveResults:%@", products);
     self.products = products;
     
     if (products && products.count > 0)
@@ -208,7 +323,7 @@
 
 - (void)sfCameraViewController:(SFCameraViewController *)cameraViewController didDetectBarcode:(SFBarcode *)barcode
 {
-    NSLog(@"Recognized Barcode Text = %@", barcode.text);
+    NSLog(@"sfCameraViewController:didDetectBarcode:%@", barcode.text);
     
     dispatch_async(dispatch_get_main_queue(), ^{
         UIAlertView *alert = [[UIAlertView alloc] initWithTitle:barcode.typeString message:barcode.text delegate:self cancelButtonTitle:@"OK" otherButtonTitles:nil];
@@ -228,12 +343,33 @@
 #pragma mark -
 #pragma mark IBActions
 
+- (IBAction)togglePremiumPublic:(UIButton *)sender
+{
+    if (_clientMode == ClientModePremium)
+        self.clientMode = ClientModePublic;
+    else if (_clientMode == ClientModePublic)
+        self.clientMode = ClientModePremium;
+}
+
+
 - (IBAction)acceptButtonClicked:(id)sender
 {
     [_clientIDTextField resignFirstResponder];
+    [_appIDTextField resignFirstResponder];
+    [_appKeyTextField resignFirstResponder];
     
-    NSString *clientID = [_clientIDTextField.text stringByTrimmingCharactersInSet:[NSCharacterSet whitespaceAndNewlineCharacterSet]];
-    self.clientID = clientID;
+    if (_clientMode == ClientModePremium)
+    {
+        NSString *clientID = [_clientIDTextField.text stringByTrimmingCharactersInSet:[NSCharacterSet whitespaceAndNewlineCharacterSet]];
+        self.clientID = clientID;
+    }
+    else if (_clientMode == ClientModePublic)
+    {
+        NSString *appID = [_appIDTextField.text stringByTrimmingCharactersInSet:[NSCharacterSet whitespaceAndNewlineCharacterSet]];
+        self.appID = appID;
+        NSString *appKey = [_appKeyTextField.text stringByTrimmingCharactersInSet:[NSCharacterSet whitespaceAndNewlineCharacterSet]];
+        self.appKey = appKey;
+    }
     
     self.hud.labelText = @"Verifying...";
     self.hud.mode = MBProgressHUDModeIndeterminate;
@@ -241,35 +377,78 @@
     
     self.slyce = [SFSlyce sharedInstance];
     dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), ^{
+        
         NSError *error = nil;
-        BOOL success = [_slyce openWithClientID:self.clientID error:&error];
-        if (!success)
+        if (_clientMode == ClientModePremium)
         {
-            dispatch_async(dispatch_get_main_queue(), ^{
-                [self.hud hide:YES];
-                self.premiumLabel.text = [error sf_message];
-            });
+            BOOL success = [_slyce openWithClientID:self.clientID error:&error];
+            if (!success)
+            {
+                dispatch_async(dispatch_get_main_queue(), ^{
+                    [self.hud hide:YES];
+                    self.premiumLabel.text = [error sf_message];
+                });
+            }
+            else
+            {
+                dispatch_async(dispatch_get_main_queue(), ^{
+                    [self.hud hide:YES];
+                    NSString *isPremium = _slyce.isPremiumUser ? @"YES" : @"NO";
+                    NSString *is2DEnabled = _slyce.is2DSearchEnabled ? @"YES" : @"NO";
+                    self.premiumLabel.text = [NSString stringWithFormat:@"Premium = %@, 2D Enabled = %@", isPremium, is2DEnabled];
+                });
+            }
         }
-        else
+        else if (_clientMode == ClientModePublic)
         {
-            dispatch_async(dispatch_get_main_queue(), ^{
-                [self.hud hide:YES];
-                NSString *isPremium = _slyce.isPremiumUser ? @"YES" : @"NO";
-                NSString *is2DEnabled = _slyce.is2DSearchEnabled ? @"YES" : @"NO";
-                self.premiumLabel.text = [NSString stringWithFormat:@"Premium = %@, 2D Enabled = %@", isPremium, is2DEnabled];
-            });
+            BOOL success = [_slyce openWithAppId:self.appID appKey:self.appKey error:&error];
+            if (!success)
+            {
+                dispatch_async(dispatch_get_main_queue(), ^{
+                    [self.hud hide:YES];
+                    self.publicLabel.text = [error sf_message];
+                });
+            }
+            else
+            {
+                dispatch_async(dispatch_get_main_queue(), ^{
+                    [self.hud hide:YES];
+                    self.publicLabel.text = @"Success";
+                });
+            }
         }
+        
     });
 }
 
 //Headless mode
 - (IBAction)recognizeImage:(UIButton *)sender
 {
-    if (!_clientID || [_clientID isEqualToString:@""])
+    if (_clientMode == ClientModePremium)
     {
-        [[[UIAlertView alloc] initWithTitle:@"Error" message:@"Please enter a client ID" delegate:nil cancelButtonTitle:@"OK" otherButtonTitles:nil] show];
+        if (!self.clientID || [self.clientID isEqualToString:@""])
+        {
+            [[[UIAlertView alloc] initWithTitle:@"Error" message:@"Please enter a client ID" delegate:nil cancelButtonTitle:@"OK" otherButtonTitles:nil] show];
+            
+            return;
+        }
+    }
+    
+    else if (_clientMode == ClientModePublic)
+    {
+        if (!self.appID || [self.appID isEqualToString:@""])
+        {
+            [[[UIAlertView alloc] initWithTitle:@"Error" message:@"Please enter an app ID" delegate:nil cancelButtonTitle:@"OK" otherButtonTitles:nil] show];
+            
+            return;
+        }
         
-        return;
+        if (!self.appKey || [self.appKey isEqualToString:@""])
+        {
+            [[[UIAlertView alloc] initWithTitle:@"Error" message:@"Please enter an app Key" delegate:nil cancelButtonTitle:@"OK" otherButtonTitles:nil] show];
+            
+            return;
+        }
     }
     
     UIImage *img = [UIImage imageNamed:@"shoe.jpg"];
@@ -284,16 +463,37 @@
     
     SFRequest *request = [[SFRequest alloc] initWithSlyce:self.slyce options:nil andDelegate:self];
     [request getProductsFromImage:img merchantIDs:nil];
+    //[request getItemDescriptionFromImage:img];
 }
 
 //Camera/Headless mode
 - (IBAction)openCustomCameraViewController:(UIButton *)sender
 {
-    if (!_clientID || [_clientID isEqualToString:@""])
+    if (_clientMode == ClientModePremium)
     {
-        [[[UIAlertView alloc] initWithTitle:@"Error" message:@"Please enter a client ID" delegate:nil cancelButtonTitle:@"OK" otherButtonTitles:nil] show];
+        if (!self.clientID || [self.clientID isEqualToString:@""])
+        {
+            [[[UIAlertView alloc] initWithTitle:@"Error" message:@"Please enter a client ID" delegate:nil cancelButtonTitle:@"OK" otherButtonTitles:nil] show];
+            
+            return;
+        }
+    }
+    
+    else if (_clientMode == ClientModePublic)
+    {
+        if (!self.appID || [self.appID isEqualToString:@""])
+        {
+            [[[UIAlertView alloc] initWithTitle:@"Error" message:@"Please enter an app ID" delegate:nil cancelButtonTitle:@"OK" otherButtonTitles:nil] show];
+            
+            return;
+        }
         
-        return;
+        if (!self.appKey || [self.appKey isEqualToString:@""])
+        {
+            [[[UIAlertView alloc] initWithTitle:@"Error" message:@"Please enter an app Key" delegate:nil cancelButtonTitle:@"OK" otherButtonTitles:nil] show];
+            
+            return;
+        }
     }
     
     [self performSegueWithIdentifier:@"MyCameraViewControllerSegue" sender:_slyce];
@@ -302,14 +502,35 @@
 //Full UI
 - (IBAction)openSlyceCameraViewController:(UIButton *)sender
 {
-    if (!_clientID || [_clientID isEqualToString:@""])
+    if (_clientMode == ClientModePremium)
     {
-        [[[UIAlertView alloc] initWithTitle:@"Error" message:@"Please enter a client ID" delegate:nil cancelButtonTitle:@"OK" otherButtonTitles:nil] show];
+        if (!self.clientID || [self.clientID isEqualToString:@""])
+        {
+            [[[UIAlertView alloc] initWithTitle:@"Error" message:@"Please enter a client ID" delegate:nil cancelButtonTitle:@"OK" otherButtonTitles:nil] show];
+            
+            return;
+        }
+    }
+    
+    else if (_clientMode == ClientModePublic)
+    {
+        if (!self.appID || [self.appID isEqualToString:@""])
+        {
+            [[[UIAlertView alloc] initWithTitle:@"Error" message:@"Please enter an app ID" delegate:nil cancelButtonTitle:@"OK" otherButtonTitles:nil] show];
+            
+            return;
+        }
         
-        return;
+        if (!self.appKey || [self.appKey isEqualToString:@""])
+        {
+            [[[UIAlertView alloc] initWithTitle:@"Error" message:@"Please enter an app Key" delegate:nil cancelButtonTitle:@"OK" otherButtonTitles:nil] show];
+            
+            return;
+        }
     }
     
     self.cameraVC = [[SFCameraViewController alloc] initWithSlyce:_slyce resourcesBundle:nil options:nil andDelegate:self];
+    //self.cameraView.shouldUseContinuousRecognition = NO;
     [_cameraVC presentFromViewController:self usingAnimation:SFAnimationTypeZoom completionBlock:^{
         NSLog(@"SFCameraViewController was presented");
     }];
@@ -324,17 +545,23 @@
 {
     [super viewDidLoad];
     
+    if ([[NSUserDefaults standardUserDefaults] objectForKey:kClientMode])
+    {
+        NSString *mode = [[NSUserDefaults standardUserDefaults] objectForKey:kClientMode];
+        if ([mode isEqualToString:@"premium"])
+            self.clientMode = ClientModePremium;
+        else if ([mode isEqualToString:@"public"])
+            self.clientMode = ClientModePublic;
+    }
+    else
+        self.clientMode = ClientModePremium;
+    
     self.hud = [[MBProgressHUD alloc] initWithView:self.view];
     self.hud.dimBackground = YES;
     self.hud.animationType = MBProgressHUDAnimationFade;
     [self.view addSubview:self.hud];
     
     self.versionLabel.text = [NSString stringWithFormat:@"v%@", [[[NSBundle mainBundle] infoDictionary] objectForKey:@"CFBundleShortVersionString"]];
-    self.clientIDTextField.text = [[NSUserDefaults standardUserDefaults] objectForKey:kClientID];
-    _clientID = self.clientIDTextField.text;
-    
-    if (_clientID && ! [_clientIDTextField.text isEqualToString:@""])
-        [self acceptButtonClicked:nil];
 }
 
 @end